Manually update all ECMWF code
