+++ packages/contracts/src/contracts/geo_schemas.py
@@ -11,8 +11,9 @@ class H3GridWeights(pt.Model):
     """
 
     h3_index: int = pt.Field(dtype=pl.UInt64)
-    nwp_lat: float = pt.Field(dtype=pl.Float64, ge=-90, le=90)
-    nwp_lng: float = pt.Field(dtype=pl.Float64, ge=-180, le=180)
-    len: int = pt.Field(dtype=pl.UInt32)
-    total: int = pt.Field(dtype=pl.UInt32)
-    proportion: float = pt.Field(dtype=pl.Float64)
+    nwp_lat: float = pt.Field(dtype=pl.Float32, ge=-90, le=90)
+    nwp_lon: float = pt.Field(dtype=pl.Float32, ge=-180, le=180)
+    proportion: float = pt.Field(
+        dtype=pl.Float32,
+        description="The proportion of this NWP grid box overlaps with the H3 hexagon",
+    )
