+++ packages/contracts/src/contracts/settings.py
@@ -84,16 +84,6 @@ class Settings(BaseSettings):
             },
         )
 
-    # ECMWF source bucket.
-    ecmwf_s3_bucket: str = Field(
-        default="dynamical-ecmwf-ifs-ens",
-        description="S3 bucket for ECMWF Icechunk store.",
-    )
-    ecmwf_s3_prefix: str = Field(
-        default="ecmwf-ifs-ens-forecast-15-day-0-25-degree/v0.1.0.icechunk/",
-        description="S3 prefix for ECMWF Icechunk store.",
-    )
-
     # Paths to the data we manage
     nged_data_path: Path = Path("data/NGED")
     nwp_data_path: Path = Path("data/NWP")
