--- packages/dynamical_data/src/dynamical_data/__init__.py
@@ -1,4 +1,0 @@
-from .processing import download_and_scale_ecmwf
-from .scaling import recover_physical_units
-
-__all__ = [download_and_scale_ecmwf, recover_physical_units]
