--- packages/dynamical_data/src/dynamical_data/scripts/__init__.py
@@ -1,4 +1,0 @@
-"""Scripts for testing and development.
-
-This package contains test data generation scripts and other development tools.
-"""
